+++ pkg/handlers/platform/chart/suite_test.go
@@ -33,12 +33,6 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/operatorV2/operation"
 )
 
-//go:embed suite/example-1.0.0.tgz
-var chart_1_0 []byte
-
-//go:embed suite/example-1.1.0.tgz
-var chart_1_1 []byte
-
 func newFakeHandler() *handler {
 	f := fakeClientSet.NewSimpleClientset()
 	k := fake.NewSimpleClientset()